requirements.txt
Flask 
requests
protobuf
protobuf_decoder
pycryptodome
httpx
